requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.27,<0.3.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.20.0